+++ watch/mlops/schedule_evaluation.py
@@ -271,8 +271,16 @@ def schedule_evaluation(config):
     dag.print_graphs()
     dag.inspect_configurables()
 
-    queue_dpath = root_dpath / '_cmd_queue_schedule'
-    queue_dpath.ensuredir()
+    if config.run:
+        mlops_meta = (root_dpath / '_mlops_schedule').ensuredir()
+        (root_dpath / '_cmd_queue_schedule').ensuredir()
+        # Write some metadata to help aggregate set its defaults automatically
+        most_recent_fpath = mlops_meta / 'most_recent_run.json'
+        data = {
+            'pipeline': str(config.pipeline),
+        }
+        import json
+        most_recent_fpath.write_text(json.dumps(data, indent='    '))
 
     queue = config.create_queue(gpus=config.devices)